remove stacks.index, optimize organizationGroups component
