--- app/stacks/index/template.hbs
@@ -1,1 +1,0 @@
-{{stacks-by-organization stacks=model}}
