+++ app/stacks/template.hbs
@@ -1,1 +1,1 @@
-{{outlet}}
+{{stacks-by-organization stacks=model}}
